create model folder
